requirements.txt
nltk==3.2
python-docx==0.8.5
PyNLPl==0.8.0
